// platform/frontend/src/client/types.gen.ts
export type ChapterCreate = {
    /**
     * Title
     */
    title: string;
    /**
     * Description
     */
    description?: string | null;
    /**
     * Order Index
     */
    order_index?: number;
};